feat: use c to provide memory and string operation
