+++ boot/loader.c
@@ -1,6 +1,7 @@
 #include "asm.h"
 #include "loader.h"
 #include "elf.h"
+#include "klib.h"
 
 #define SECTSIZE 512
 #define CLUSSIZE 512
@@ -17,14 +18,6 @@ static void load_kernel(u32 start_clus, void *kernel_addr);
 
 static inline void itohstr(int num, char *buf);
 static inline int strncmp(const char *s1, const char *s2, size_t n);
-
-static void memset(void *dst, u8 c, size_t siz) {
-    // TODO: Optimize performance
-    for(int i = 0; i < siz; i++) {
-        *(u8*)dst = c;
-    }
-}
-void* memcpy(void *dst, void *src, size_t siz);
 
 __attribute__((noreturn)) void main() {
     char s[] = {'Y', 0xF, 'e', 0xF, 's', 0xF};
@@ -142,30 +135,5 @@ static void load_kernel(u32 start_clus, void *kernel_addr) {
             memcpy(pa, kernel_addr + ph->off, ph->filesz);
             memset(pa + ph->filesz, 0, ph->memsz - ph->filesz);
         }
-    }
-}
-
-
-static inline void itohstr(int num, char *buf) {
-    for(int i = 15; i >= 0; i -= 2) {
-        buf[i] = 0xF;
-        buf[i-1] = num % 16;
-        buf[i-1] += buf[i-1] > 9 ? 'A' - 10 : '0';
-        num /= 16;
-    }
-    buf[16] = '$';
-    buf[17] = 0xF;
-}
-
-static inline int strncmp(const char *s1, const char *s2, size_t n) {
-    int i_s1 = 0, i_s2 = 0;
-    while (n && s1[i_s1] != '\0' && s2[i_s2] != '\0' && s1[i_s1] == s2[i_s2]) {
-        --n; ++i_s1; ++i_s2;
-    }
-    if (n == 0 || s1[i_s1] == s2[i_s2]) {
-        /* Same str */
-        return 0;
-    } else {
-        return -1;
     }
 }
